# Conflicts:
#	build.gradle
#	src/main/java/com/rest/springbootemployee/controller/EmployeeController.java
#	src/main/java/com/rest/springbootemployee/entity/Company.java
#	src/main/java/com/rest/springbootemployee/service/CompanyService.java
#	src/main/java/com/rest/springbootemployee/service/EmployeeService.java
#	src/main/resources/application.yml
#	src/test/java/com/rest/springbootemployee/EncryptPwdTest.java
#	src/test/java/com/rest/springbootemployee/companyTest/controllerTests/SpringBootCompanyApplicationTests.java
#	src/test/java/com/rest/springbootemployee/employeeTests/controllerTests/SpringBootEmployeeApplicationTests.java
#	src/test/java/com/rest/springbootemployee/employeeTests/serviceTests/EmployeeServiceTests.java
#	src/test/resources/application-test.yml

+++ src/main/java/com/rest/springbootemployee/service/CompanyService.java
@@ -12,8 +12,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
+
 
 
 @Service
@@ -53,6 +52,7 @@ public class CompanyService {
     }
 
     public void deleteCompany(Integer id) {
+        if(!companyJpaRepository.existsById(id)) throw new CompanyNotFoundException();
         companyJpaRepository.deleteById(id);
     }
 }
